// services/src/types.ts
export interface StudioLog {
    id: string;
    timestamp: string;
    type: 'COMMAND' | 'RESPONSE' | 'ERROR' | 'INFO';
    content: string;
}